move classes to their own modules
